% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{ModifiedChannelThresholds}
\alias{ModifiedChannelThresholds}
\title{Main Shiny app function call}
\usage{
ModifiedChannelThresholds(...)
}
\arguments{
\item{...}{not used}
}
\value{
shiny app object
}
\description{
Main Shiny app function call
}
